Add a new "FLY_IDLE" animation

+++ src/main/java/by/dragonsurvivalteam/dragonsurvival/client/handlers/ClientFlightHandler.java
@@ -294,7 +294,7 @@ public class ClientFlightHandler {
         }
 
         if (!Float.isFinite(lastFlightBaseFov)) {
-            lastFlightBaseFov = event.getFOV();
+            lastFlightBaseFov = (float) event.getFOV();
         }
 
         double zoom = Math.max(lastZoom, 0.01F);
